Reduce usage of TriplestoreConnection in favor of StatementSource (#298)
This refactoring changes the companion object for
`TriplestoreConnection` into one for `StatementSource`, since that is
the only interface the contained methods depend on, and updates all the
usage sites accordingly.
This leaves us in a situation where very few places actually require a
`TripleStoreConnection` (which carries much more context), and makes the
rest of the locations easier to reason about and easier to test.