WIP Keywords-fetch magic query continued
Apart from code polishing, this is a beginning of a change to have a simpler syntax in the query
(just select distinct ?keyword instead of inventing a magic function), but also allowing additional
filters on ?spec or ?station to be embedded into the query, be executed by the vanilla engine, and
then fed into the magic node by `getFilterEnrichedDobjFetch` method. The query can return the keywords
separately, this will look more SPARQL-idiomatic.