Improve parallel query test in SparqlRouteTests (#286)
In the previous implementation of the test, it is not clear that we are
getting rejections because of parallel queries, and the test runs for
longer than necessary since we are relying on a long running query.
This version instead launches multiple short queries, logs their
timestamps (so it is clear they are running in parallel when looking at
test output), and verifies that some fail and some succeed.