Enable -Yexplicit-nulls (#310)
This enables the `-Yexplicit-nulls` warning, and includes edits to make
things compile again. See
https://docs.scala-lang.org/scala3/reference/experimental/explicit-nulls.html
for reference.

The main value of this, currently, is that we can now use this warning
in any new code we write or edit, removing the `unsafeNulls` imports as
we go.

In most files, I simply added `import scala.language.unsafeNulls`,
meaning things are just as before. In some other cases I have added
`.nn` invocations, which also should keep things as they were before,
but with more explicitly visible locations where null-exceptions may
trigger at runtime (as before, but now explicit). Some cases are
false-positives, specifically in relation to string operations. See
https://stackoverflow.com/questions/70747784/scala-3-explicit-nulls-flag-makes-string-operations-quite-unusable
That may be possible to improve with a higher scala version, or with
flexible types
(https://docs.scala-lang.org/scala3/reference/experimental/explicit-nulls.html#java-interoperability-and-flexible-types-1),
but I did not look closer at that yet.

There is still one remaining issue to make this work. `sbt-buildinfo`
generates one line which fails with `explicit-nulls`. I did not find a
way around that yet, but I believe some edits could be done to
`icoscp-sbt-deploy` to configure `sbt-buildinfo` and make it behave
better.